% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hfr.R
\name{plot.hfr}
\alias{plot.hfr}
\title{Plot the dendrogram of an HFR model}
\usage{
\method{plot}{hfr}(object)
}
\arguments{
\item{object}{Fitted 'hfr' model.}
}
\value{
A plotted dendrogram.
}
\description{
Plots the dendrogram of a fitted \code{hfr} model. The heights of the
dendrogram are given by a shrinkage vector, with a maximum (unregularized) height
of \eqn{(p-1)}{(p - 1)}. Stronger shrinkage leads to a shallower hierarchy.
}
\details{

}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = hfr(x, y, factors = 0.5)
plot(fit)

}
\references{

}
\seealso{
\code{hfr}, \code{predict} and \code{coef} methods
}
\author{
Johann Pfitzinger
}
