% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hfr.R
\name{cv.hfr}
\alias{cv.hfr}
\title{Cross validation for a Hierarchical Feature Regression}
\usage{
cv.hfr(
  x,
  y,
  penalty_grid = NULL,
  factors_grid = NULL,
  q = NULL,
  intercept = TRUE,
  standardize = TRUE,
  cluster_method = c("complete", "single", "average", "ward.D2", "mcquitty", "median",
    "centroid")
)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{penalty_grid}{A vector of penalties on the effective degrees of freedom of the regression.}

\item{factors_grid}{A vector of target effective degrees of freedom of the regression.}

\item{q}{The quantile cut-off (in terms of information contributed) above which to consider levels in the hierarchy.}

\item{intercept}{Should intercept be fitted (default=TRUE).}

\item{standardize}{Logical flag for x variable standardization prior to fitting the model. The coefficients are always returned on the original scale. Default is \code{standardize=TRUE}.}

\item{cluster_method}{hierarchical cluster algorithm used to construct an asset hierarchy.}
}
\value{
An 'hfr' regression object.
}
\description{
The hierarchical feature regression is fitted by estimating a semi-supervised
hierarchical graph based on bi-variate partial correlations, decomposing a least
squares estimate along the estimated hierarchy and shrinking coefficients along the branches of the tree.
}
\details{
This function fits an HFR to a grid of penalty or factor values. The results is a
matrix of coefficients with one column for each hyperparameter. By evaluating all hyperparameters
in a single function, the speed of the algorithm can be improved substantially (e.g. by estimating
level-specific regressions only once).
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = cv.hfr(x, y, factors_grid = seq(0, 1, by = 0.1))
coef(fit)

}
\references{
Pfitzinger, J. (2021).
Cluster Regularization via a Hierarchical Feature Regression.
_arXiv [statML] 2107.04831_
}
\seealso{
\code{hfr}, \code{coef} and \code{predict} methods
}
\author{
Johann Pfitzinger
}
