% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr.R
\name{hfr}
\alias{hfr}
\title{Fit a Hierarchical Feature Regression}
\usage{
hfr(
  x,
  y,
  penalty = NULL,
  factors = NULL,
  q = NULL,
  intercept = TRUE,
  standardize = TRUE,
  cluster_method = c("DIANA", "single", "complete", "average", "ward")
)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.}

\item{y}{Response variable.}

\item{penalty}{A penalty on the effective degrees of freedom of the regression.}

\item{factors}{The target effective degrees of freedom of the regression.}

\item{q}{The quantile cut-off (in terms of information contributed) above which to consider levels in the hierarchy.}

\item{intercept}{Should intercept be fitted (default=TRUE).}

\item{standardize}{Logical flag for x variable standardization prior to fitting the model.}

\item{cluster_method}{hierarchical cluster algorithm used to construct an asset hierarchy.}
}
\value{
An 'hfr' regression object.
}
\description{
The hierarchical feature regression is fitted by estimating a semi-supervised
hierarchical graph based on bi-variate partial correlations, decomposing a least
squares estimate along the estimated hierarchy and shrinking coefficients along the branches of the tree.
}
\details{
Shrinkage can be imposed using two mechanisms: using a penalty on the effective degrees of
freedom of the regression, or targetting an explicit effective degrees of freedom.
Setting the argument \code{penalty} to a positive value, implements the former approach, while
setting the argument \code{factors} to a value between 0 and 1 implements the latter.
If neither \code{penalty} or \code{factors} is set, a linear regression with \code{penalty = 0} is
estimated.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.

For high-dimensional problems, the hierarchy becomes very large. Setting \code{q} to a value below 1
reduces the number of levels used in the hierarchy. \code{q} represents a quantile-cutoff of amount of
information contributed by the levels. The default (\code{q = 1}) considers all levels.
}
\examples{
# Load returns of assets or portfolios
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit1 = glmnet(x, y)
print(fit1)

}
\references{

}
\author{
Johann Pfitzinger
}
