% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hfr.R
\name{plot.hfr}
\alias{plot.hfr}
\title{Plot the dendrogram of an HFR model}
\usage{
\method{plot}{hfr}(x, show_details = TRUE, confidence_level = 0, max_leaf_size = 3, ...)
}
\arguments{
\item{x}{Fitted 'hfr' model.}

\item{show_details}{print model details on the plot.}

\item{confidence_level}{coefficients with a lower approximate statistical confidence are highlighted in the plot, see details. Default is \code{confidence_level=0}.}

\item{max_leaf_size}{maximum size of the leaf nodes. Default is \code{max_leaf_size=3}.}

\item{...}{additional methods passed to \code{plot}.}
}
\value{
A plotted dendrogram.
}
\description{
Plots the dendrogram of a fitted \code{hfr} model. The heights of the
levels in the dendrogram are given by a shrinkage vector, with a maximum (unregularized)
overall graph height of \eqn{p}{p} (the number of covariates in the regression).
Stronger shrinkage leads to a shallower hierarchy.
}
\details{
The dendrogram is generated using hierarchical clustering and modified
so that the height differential between any two splits is the shrinkage weight of
the lower split (ranging between \code{0} and \code{1}). With no shrinkage, all shrinkage weights
are equal to \code{1} and the dendrogram has a height of \eqn{p}{p}. With shrinkage
the dendrogram has a height of \eqn{(\kappa \times p)}{(\code{kappa} x p)}.

The leaf nodes are colored to indicate the coefficient sign, with the size indicating
the absolute magnitude of the coefficients.

The average standard errors along the branch of each coefficient can be used
to highlight coefficients that are not statistically significant. When
\code{confidence_level > 0}, branches with a lower confidence are plotted
as dotted lines.

A color bar on the right indicates the relative contribution of each level to the
coefficient of determination, with darker hues representing a larger contribution.
}
\examples{
x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit = hfr(x, y, kappa = 0.5)
plot(fit)

}
\seealso{
\code{\link{hfr}}, \code{\link{se.avg}}, \code{\link{predict}} and \code{\link{coef}} methods
}
\author{
Johann Pfitzinger
}
